PreArtDropout events.
PreArtTest Dropout
PreArtResult Dropout.

The effect of these two events is that it off sets dropout from
pre-ART care.

Before: if you were not going to get a CD4 test result then you were
immediately lost. Even though you wouldn't be classed as lost
(in reality) by the clinic until you had 'missed' an appointment.

Now: the model schedules a dropout event later in time, depending
upon if the event was a CD4 test result or a CD4 test, so a patient
is not lost until they have effectively 'missed' the appointment.

This should not alter results substantially but is an itch that has
now been scratched.