Calibration on 2nd April. RESULT.
Calibration fits both the 'input parameters' and 'calibration points' as best as possible.
With time dependent linkage on HBCT, we allow a roll-out period of low linkage to care then scale that up highly.
This allows linkage from VCT / PICT to be much higher and still achieve the correct distribution of individuals entering care.
Before, I had to use low PICT / VCT linkage rates to allow HBCT to have an impact in C1, as HBCT linkage was very poor.

Now, linkage from PICT = 0.95 and VCT = 0.9. With VCT linkage at 0.7 we fit the perfect distribution in period 1 of C1 but sacrifice # initiating ART in this tab. As a result, linkage rates are kept high in order to achieve a better fit to the two figures on this page.
Additionally, for short/long term retention in pre-ART care, all values are x / 0.8 to account for the 20% of patients that will not attend their appointment.
In addition, probability attending secondary CD4 test (R2), we assume the denominator is too large here as a proportion will be eligible for treatment after 1st CD4 test so will not require a secondary test. From L2.1 we know that roughly 20% of patients are eligible at first CD4 test, so we assume that the denominator is 20% too large and adjust all values by x / 0.8 too account for this.